% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit the matrix-based model with Stan}
\usage{
fit_model(parameters, m, ...)
}
\arguments{
\item{parameters}{A list from \code{model_parameters()}.}

\item{m}{An integer scalar >= to the number of column in \code{d}.}

\item{...}{Other arguments passed on to \code{rstan::sampling} (e.g. iter, chains,
etc.).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
\code{fit_model} returns a stanfit object. The \code{fit_model}
function fit the matrix \code{d} to the matrix-based model for a given
number of column \code{m} and and prior hyper-parameters for the
logit-normal distribution (mean and variance)
\describe{ \item{mu}{A Gaussian distribution with mean \code{mu_prior_mean}
and standard deviation \code{mu_prior_sd} parameters} \item{s2}{An inverse
gamma distribution with shape \code{s2_prior_a} and scale
\code{s2_prior_b}}}
}
\seealso{
\url{https://mc-stan.org/math/}.

\code{\link[=model_parameters]{model_parameters()}}.

Other fitting functions: 
\code{\link{integrated_likelihood}()},
\code{\link{posterior_analysis}()},
\code{\link{posterior_mu_s2}()},
\code{\link{posterior_pl}()}
}
\concept{fitting functions}
