% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{integrated_likelihood}
\alias{integrated_likelihood}
\title{Compute the integrated likelihood}
\usage{
integrated_likelihood(stan_fit, seed = NULL, ...)
}
\arguments{
\item{stan_fit}{A stanfit object.}

\item{seed}{A integer scalar. If \code{NULL}, no seed is set for bridge
sampling.}

\item{...}{Other arguments passed on to \code{bridgesampling::bridge_sampler}.}
}
\value{
A numeric scalar.
}
\description{
\code{integrated_likelihood} compute the integrated log
marginal likelihood via bridge sampling.
}
\seealso{
Other fitting functions: 
\code{\link{fit_model}()},
\code{\link{posterior_analysis}()},
\code{\link{posterior_mu_s2}()},
\code{\link{posterior_pl}()}
}
\concept{fitting functions}
