% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters.R
\name{model_parameters}
\alias{model_parameters}
\title{List of parameters for fitting the matrix-based model}
\usage{
model_parameters(
  d,
  mu_prior_mean = 0,
  mu_prior_sd = 1.5,
  s2_prior_a = 20,
  s2_prior_b = 20
)
}
\arguments{
\item{d}{A logical matrix.}

\item{mu_prior_mean}{A numeric scalar. The default is 0.}

\item{mu_prior_sd}{A numeric scalar >0. The default is 1.5.}

\item{s2_prior_a}{A numeric scalar >0. The default is 20.}

\item{s2_prior_b}{A numeric scalar >0. The default is 20.}
}
\value{
A list of parameters to be passed to fitting functions.
}
\description{
\code{discovery_parameters} returns a list of parameters used by
fitting function (e.g. \code{fit_model}) and prior hyper-parameters for the
logit-normal distribution (mean and variance) \describe{\item{mean (mu)}{A
Gaussian distribution with mean \code{mu_prior_mean} and standard deviation
\code{mu_prior_sd} parameters} \item{variance (s2)}{An inverse gamma
distribution with shape \code{s2_prior_a} and scale \code{s2_prior_b}}}.
Non-informative prior value are set as default.
}
\examples{

}
\seealso{
\code{\link[=fit_model]{fit_model()}}.
}
