% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_analysis.R
\name{multinom_expectation}
\alias{multinom_expectation}
\title{TO DO multinom_sample}
\usage{
multinom_expectation(prepost_pl_nc, prepost_pl_c, N)
}
\arguments{
\item{prepost_pl_nc}{TO DO}

\item{prepost_pl_c}{TO DO}

\item{N}{TO DO}
}
\value{

}
\description{
TO DO multinom_sample
}
