% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{posterior_analysis}
\alias{posterior_analysis}
\title{Fit the model and provide posterior sample and integrated likelihood}
\usage{
posterior_analysis(parameters, m_max, ...)
}
\arguments{
\item{parameters}{A list from \code{model_parameters()}.}

\item{m_max}{An integer scalar >= to the number of column in \code{d}.}

\item{...}{Other arguments passed on to \code{rstan::sampling} (e.g. iter, chains,
etc.).#'}
}
\value{

}
\description{
\code{posterior_analysis} fit the model using \code{fit_model}
on a m grid from 0 to \code{m_max()}. It provides \describe{\item{the
integrated likelihood}{computed via bridge sampling by
\code{integrated_likelihood}} \item{the posterior_pl}{extracted from the
MCMC chains by \code{posterior_pl}} \item{a summary of model
diagnostic}{computed by \code{fit_model}}}
}
\examples{

}
\seealso{
\code{\link[=fit_model]{fit_model()}}.

\code{\link[=integrated_likelihood]{integrated_likelihood()}}.

\code{\link[=posterior_pl]{posterior_pl()}}.

Other fitting functions: 
\code{\link{fit_model}()},
\code{\link{integrated_likelihood}()},
\code{\link{posterior_mu_s2}()},
\code{\link{posterior_pl}()}
}
\concept{fitting functions}
