% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{posterior_mu_s2}
\alias{posterior_mu_s2}
\title{Extract posterior sample for mu and s2}
\usage{
posterior_mu_s2(stan_fit)
}
\arguments{
\item{stan_fit}{A stanfit object.}
}
\value{
A numeric scalar.
}
\description{
\code{posterior_mu_s2} extract the posterior sample of mu and s2 from the
MCMC chains
}
\seealso{
\url{https://mc-stan.org/math/}.

Other fitting functions: 
\code{\link{fit_model}()},
\code{\link{integrated_likelihood}()},
\code{\link{posterior_analysis}()},
\code{\link{posterior_pl}()}
}
\concept{fitting functions}
