% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{posterior_pl}
\alias{posterior_pl}
\title{Extract posterior sample for pl}
\usage{
posterior_pl(stan_fit)
}
\arguments{
\item{stan_fit}{A stanfit object.}
}
\value{
A numeric scalar.
}
\description{
\code{posterior_pl} extract the posterior sample of pl from the
MCMC chains
}
\seealso{
\url{https://mc-stan.org/math/}.

Other fitting functions: 
\code{\link{fit_model}()},
\code{\link{integrated_likelihood}()},
\code{\link{posterior_analysis}()}
}
\concept{fitting functions}
