% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_analysis.R
\name{prepost_analysis}
\alias{prepost_analysis}
\title{TO DO prepost_analysis}
\usage{
prepost_analysis(fit_nc, fit_c, sample_size, n_end_users, sim)
}
\arguments{
\item{fit_nc}{TO DO}

\item{fit_c}{TO DO}

\item{sample_size}{TO DO}

\item{n_end_users}{TO DO}

\item{sim}{TO DO}
}
\value{

}
\description{
TO DO prepost_analysis
}
