% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_analysis.R
\name{prepost_pl}
\alias{prepost_pl}
\title{TO DO prepost_pl}
\usage{
prepost_pl(m, pl, n, seed = NULL)
}
\arguments{
\item{m}{TO DO}

\item{pl}{TO DO}

\item{n}{TO DO}

\item{seed}{TO DO}
}
\value{

}
\description{
TO DO prepost_pl
}
