% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_analysis.R
\name{rstan_p}
\alias{rstan_p}
\title{Title TO DO}
\usage{
rstan_p(nb_sample, n, mu, s2, ...)
}
\arguments{
\item{n}{TO DO}

\item{mu}{TO DO}

\item{s2}{TO DO}
}
\value{
TO DO
}
\description{
Title TO DO
}
