% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_analysis.R
\name{sample_posterior_pl}
\alias{sample_posterior_pl}
\title{TO DO sample_posterior_pl}
\usage{
sample_posterior_pl(nb_sample, n, mu, s2, sampler = "rstan")
}
\arguments{
\item{nb_sample}{TO DO}

\item{n}{TO DO}

\item{mu}{TO DO}

\item{s2}{TO DO}

\item{metropolis}{TO DO}
}
\value{

}
\description{
TO DO sample_posterior_pl
}
