% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepost_analysis.R
\name{prepost_pl}
\alias{prepost_pl}
\title{TO DO prepost_pl}
\usage{
prepost_pl(m, pl, mu_s2, n, n_init, seed = NULL)
}
\arguments{
\item{m}{a vector}

\item{pl}{a list}

\item{n}{a vector}

\item{seed}{a number for the seed, default = NULL}
}
\value{

}
\description{
TO DO prepost_pl
}
